requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
